+++ NVs.OccupancySensor.CV.Tests/OccupancySensorShould.cs
@@ -70,8 +70,8 @@ namespace NVs.OccupancySensor.CV.Tests
             camera.SetupGet(c => c.IsRunning).Returns(true);
             camera.SetupGet(c => c.Stream).Returns(new CameraStream(capture.Object, CancellationToken.None, new Mock<ILogger<CameraStream>>().Object, TimeSpan.FromMilliseconds(100)));
             
-            converter.Setup(c => c.Convert(It.IsAny<Mat>())).Returns(() => new Image<Rgb, int>(100,100)).Verifiable("Convert was not called!");
-            detector.Setup(d => d.Detect(It.IsAny<Image<Rgb,int>>())).Returns<Image<Rgb,int>>(x => x).Verifiable("Detect was not called!");
+            converter.Setup(c => c.Convert(It.IsAny<Mat>())).Returns(() => new Image<Rgb, float>(100,100)).Verifiable("Convert was not called!");
+            detector.Setup(d => d.Detect(It.IsAny<Image<Rgb, float>>())).Returns<Image<Rgb, float>>(x => x).Verifiable("Detect was not called!");
             var sensor = new Impl.OccupancySensor(camera.Object, converter.Object, detector.Object, logger.Object);
             
             camera.Raise(c => c.PropertyChanged += null, new PropertyChangedEventArgs(nameof(ICamera.IsRunning)));
